ref(api): Split `CurlConnectionManager` into its own submodule
